requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiosqlite
pyjwt
passlib
bcrypt